Refactor Creator, Locator, Mutator, and Remover into wildfly-clustering-ee-spi.
