+++ clustering/ee/spi/src/main/java/org/wildfly/clustering/ee/Remover.java
@@ -19,7 +19,7 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
-package org.wildfly.clustering.ee.infinispan;
+package org.wildfly.clustering.ee;
 
 /**
  * Removes an entry from the cache
