+++ clustering/ejb/infinispan/src/main/java/org/wildfly/clustering/ejb/infinispan/BeanFactory.java
@@ -21,8 +21,8 @@
  */
 package org.wildfly.clustering.ejb.infinispan;
 
-import org.wildfly.clustering.ee.infinispan.Creator;
-import org.wildfly.clustering.ee.infinispan.Locator;
+import org.wildfly.clustering.ee.Creator;
+import org.wildfly.clustering.ee.Locator;
 import org.wildfly.clustering.ejb.Bean;
 
 /**
