+++ clustering/ejb/infinispan/src/test/java/org/wildfly/clustering/ejb/infinispan/bean/InfinispanBeanTestCase.java
@@ -35,7 +35,7 @@ import org.junit.After;
 import org.junit.Assert;
 import org.junit.Test;
 import org.mockito.Matchers;
-import org.wildfly.clustering.ee.infinispan.Mutator;
+import org.wildfly.clustering.ee.Mutator;
 import org.wildfly.clustering.ejb.Bean;
 import org.wildfly.clustering.ejb.PassivationListener;
 import org.wildfly.clustering.ejb.RemoveListener;
