+++ clustering/web/infinispan/src/main/java/org/wildfly/clustering/web/infinispan/session/InfinispanSession.java
@@ -24,7 +24,7 @@ package org.wildfly.clustering.web.infinispan.session;
 import java.time.Instant;
 import java.util.concurrent.atomic.AtomicReference;
 
-import org.wildfly.clustering.ee.infinispan.Remover;
+import org.wildfly.clustering.ee.Remover;
 import org.wildfly.clustering.web.LocalContextFactory;
 import org.wildfly.clustering.web.session.Session;
 import org.wildfly.clustering.web.session.SessionMetaData;
