+++ clustering/web/infinispan/src/main/java/org/wildfly/clustering/web/infinispan/session/SessionFactory.java
@@ -23,10 +23,10 @@ package org.wildfly.clustering.web.infinispan.session;
 
 import java.util.Map;
 
-import org.wildfly.clustering.ee.infinispan.Creator;
+import org.wildfly.clustering.ee.Creator;
+import org.wildfly.clustering.ee.Locator;
+import org.wildfly.clustering.ee.Remover;
 import org.wildfly.clustering.ee.infinispan.Evictor;
-import org.wildfly.clustering.ee.infinispan.Locator;
-import org.wildfly.clustering.ee.infinispan.Remover;
 import org.wildfly.clustering.web.session.ImmutableSession;
 import org.wildfly.clustering.web.session.ImmutableSessionAttributes;
 import org.wildfly.clustering.web.session.ImmutableSessionMetaData;
