+++ clustering/web/infinispan/src/main/java/org/wildfly/clustering/web/infinispan/sso/SessionsFactory.java
@@ -24,9 +24,9 @@ package org.wildfly.clustering.web.infinispan.sso;
 
 import java.util.Map;
 
-import org.wildfly.clustering.ee.infinispan.Creator;
-import org.wildfly.clustering.ee.infinispan.Locator;
-import org.wildfly.clustering.ee.infinispan.Remover;
+import org.wildfly.clustering.ee.Creator;
+import org.wildfly.clustering.ee.Locator;
+import org.wildfly.clustering.ee.Remover;
 import org.wildfly.clustering.web.sso.Sessions;
 
 public interface SessionsFactory<V, D> extends Creator<String, V, Void>, Locator<String, V>, Remover<String> {
