+++ clustering/web/infinispan/src/main/java/org/wildfly/clustering/web/infinispan/sso/coarse/CoarseSessions.java
@@ -25,7 +25,7 @@ import java.util.Collections;
 import java.util.Map;
 import java.util.Set;
 
-import org.wildfly.clustering.ee.infinispan.Mutator;
+import org.wildfly.clustering.ee.Mutator;
 import org.wildfly.clustering.web.sso.Sessions;
 
 public class CoarseSessions<D> implements Sessions<D> {
