+++ clustering/web/infinispan/src/test/java/org/wildfly/clustering/web/infinispan/session/InfinispanSessionTestCase.java
@@ -28,7 +28,7 @@ import java.time.Instant;
 import java.util.concurrent.atomic.AtomicReference;
 
 import org.junit.Test;
-import org.wildfly.clustering.ee.infinispan.Remover;
+import org.wildfly.clustering.ee.Remover;
 import org.wildfly.clustering.web.LocalContextFactory;
 import org.wildfly.clustering.web.session.Session;
