+++ clustering/web/infinispan/src/test/java/org/wildfly/clustering/web/infinispan/session/SessionExpirationSchedulerTestCase.java
@@ -32,7 +32,7 @@ import java.util.concurrent.TimeUnit;
 
 import org.junit.Test;
 import org.wildfly.clustering.ee.Batcher;
-import org.wildfly.clustering.ee.infinispan.Remover;
+import org.wildfly.clustering.ee.Remover;
 import org.wildfly.clustering.ee.infinispan.TransactionBatch;
 import org.wildfly.clustering.web.session.ImmutableSessionMetaData;
